backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0
PyYAML>=6.0
requests>=2.31.0